feat: support different tips on start pages
