+++ public/translations/cy.json
@@ -2,7 +2,8 @@
   "start": {
     "title": "Dod o hyd i leoedd i ailgylchu.",
     "location": {
-      "title": "What do you need to recycle?"
+      "title": "What do you need to recycle?",
+      "exploreTheMap": "Explore the map"
     },
     "about": {
       "title": "About this service",
