+++ src/pages/Start/Location.tsx
@@ -1,13 +1,24 @@
 import { useTranslation } from 'react-i18next';
-import { useParams } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
+import '@etchteam/diamond-ui/control/Button/Button';
+
+import '../../components/Wrap/Wrap';
 
 export default function StartLocationPage() {
   const { t } = useTranslation();
-  const { postcode } = useParams();
 
   return (
-    <h2>
-      {t('start.location.title')} {postcode}
-    </h2>
+    <>
+      <locator-wrap slot="main">
+        <diamond-section padding="lg">
+          <Outlet />
+        </diamond-section>
+      </locator-wrap>
+      <div slot="aside">
+        <diamond-button>
+          <button>{t('start.location.exploreTheMap')}</button>
+        </diamond-button>
+      </div>
+    </>
   );
 }
