+++ src/pages/Start/Start.tsx
@@ -1,32 +1,28 @@
-import { useTranslation } from 'react-i18next';
-import { Form, redirect, ActionFunctionArgs } from 'react-router-dom';
-import '@etchteam/diamond-ui/composition/FormGroup/FormGroup';
+import { Outlet } from 'react-router-dom';
+import '@etchteam/diamond-ui/canvas/Section/Section';
 
-import '../../components/LocationInput/LocationInput';
-import PostCodeResolver from '../../lib/PostcodeResolver';
-
-export async function startPageAction({ request }: ActionFunctionArgs) {
-  const formData = await request.formData();
-  const location = formData.get('location') as string;
-  const postcode = await PostCodeResolver.fromString(location);
-  return redirect(`/${postcode}`);
-}
+import '../../components/Wrap/Wrap';
+import '../../components/Tip/Tip';
 
 export default function StartPage() {
-  const { t } = useTranslation();
-
   return (
     <>
-      <h2>{t('start.title')}</h2>
-      <Form method="post">
-        <diamond-form-group class="diamond-spacing-bottom-md">
-          <label htmlFor="location-input">Where are you?</label>
-          <locator-location-input></locator-location-input>
-        </diamond-form-group>
-        <diamond-button width="full-width" variant="primary">
-          <button type="submit">Get started</button>
-        </diamond-button>
-      </Form>
+      <locator-wrap slot="main">
+        <diamond-section padding="lg">
+          <Outlet />
+        </diamond-section>
+      </locator-wrap>
+      <locator-tip slot="aside">
+        <locator-wrap>
+          <p>Use this service to:</p>
+          <ul>
+            <li>see your nearest places to recycle</li>
+            <li>find out how to recycle a specific item</li>
+            <li>check what you can recycle at home</li>
+          </ul>
+          <img src="/images/recycling-technology.webp" alt="" />
+        </locator-wrap>
+      </locator-tip>
     </>
   );
 }
